requirements.txt
Django==4.2.7
djangorestframework==3.14.0
python-decouple==3.8
gunicorn==21.2.0
whitenoise==6.6.0
psycopg2-binary==2.9.9
